+++ onlineconf/config.py
@@ -8,7 +8,7 @@ import cdblib
 import yaml
 
 if TYPE_CHECKING:
-    from asyncio import AbstractEventLoop
+    from asyncio import Task
 
 
 __all__ = ("Config",)
@@ -18,29 +18,27 @@ class Config:
     def __init__(
         self,
         filename: str,
-        reload_interval: Optional[int] = None,
-        loop: Optional["AbstractEventLoop"] = None,
+        reload_interval: int = 0,
     ) -> None:
         self._filename = filename
         self._reload_interval = reload_interval
         self._reload_task: Optional[asyncio.Task[None]] = None
-        self._loop = loop if loop else asyncio.get_running_loop()
 
     @classmethod
-    def read(
+    async def read(
         cls,
         filename: str,
-        reload_interval: Optional[int] = 30,
-        loop: Optional["AbstractEventLoop"] = None,
+        reload_interval: int = 0,
     ) -> "Config":
         """Read a cdb file and schedule periodic reload if needed"""
-        _config = cls(filename=filename, reload_interval=reload_interval, loop=loop)
+        _config = cls(filename=filename, reload_interval=reload_interval)
 
         with open(_config._filename, "rb") as f:
             _config.cdb = cdblib.Reader(f.read())
 
         if reload_interval:
-            _config._reload_task = _config._loop.create_task(_config._schedule_reload())
+            _config._reload_task = asyncio.create_task(_config._schedule_reload())
+            _config._reload_task.add_done_callback(_config._reload_done_callback)
         return _config
 
     def get(self, key: str) -> Union[str, Dict[str, Any]]:
@@ -132,3 +130,15 @@ class Config:
                     yield _path, f"s{value}"
                 else:
                     yield _path, f"j{value}"
+
+    @staticmethod
+    def _reload_done_callback(task: "Task[Any]"):
+        loop = asyncio.get_running_loop()
+        try:
+            exc = task.exception()
+        except asyncio.CancelledError:
+            pass
+        else:
+            if exc:
+                context = {"exception": exc}
+                loop.call_exception_handler(context)