+++ tests/test_config.py
@@ -9,61 +9,59 @@ import cdblib
 from onlineconf import Config
 
 
-class ReadConfig(unittest.TestCase):
+class ReadConfig(unittest.IsolatedAsyncioTestCase):
     cdb_writer = cdblib.Writer
 
     def setUp(self):
         _, self.cdb_filename = tempfile.mkstemp()
         self.fp = open(self.cdb_filename, "wb")
         self.writer = self.cdb_writer(self.fp)
-        self.loop = asyncio.new_event_loop()
 
     def tearDown(self):
         self.fp.close()
         os.remove(self.cdb_filename)
-        self.loop.close()
 
     def finalize_cdb(self):
         self.writer.finalize()
         self.fp.close()
 
-    def test_read_string_value(self):
+    async def test_read_string_value(self):
         key, value = "/my/service", "123"
 
         self.writer.put(key.encode(), f"s{value}".encode())
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
 
         saved_value = conf.get(key)
         self.assertIsInstance(saved_value, str)
         self.assertEqual(saved_value, value)
 
-    def test_read_json_value(self):
+    async def test_read_json_value(self):
         key, value = "/my/service", dict(key="value")
 
         self.writer.put(key.encode(), f"j{json.dumps(value)}".encode())
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
 
         saved_value = conf.get(key)
         self.assertIsInstance(saved_value, dict)
         self.assertEqual(saved_value, value)
 
-    def test_read_json_value_with_unicode(self):
+    async def test_read_json_value_with_unicode(self):
         key, value = "/my/service", dict(key="сьешь еще этих булочек")
 
         self.writer.put(key.encode(), f"j{json.dumps(value)}".encode())
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
 
         saved_value = conf.get(key)
         self.assertIsInstance(saved_value, dict)
         self.assertEqual(saved_value, value)
 
-    def test_get_items(self):
+    async def test_get_items(self):
         items = [
             (b"/my/service/param1", b"value1"),
             (b"/my/service/param2", b"value2"),
@@ -74,43 +72,41 @@ class ReadConfig(unittest.TestCase):
             self.writer.put(*item)
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
         saved_items = conf.items()
 
         self.assertEqual(saved_items, items)
 
-    def test_get_keys(self):
+    async def test_get_keys(self):
         keys = [b"/my/service/param1", b"/my/service/param2", b"/my/service/param3"]
 
         for key in keys:
             self.writer.put(key)
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
         saved_keys = conf.keys()
 
         self.assertEqual(saved_keys, keys)
 
-    def test_contains(self):
+    async def test_contains(self):
         key = b"/my/service/param1"
 
         self.writer.put(key)
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
         self.assertIn(key, conf)  # type: ignore
         self.assertNotIn(b"nonexistent_key", conf)  # type: ignore
 
-    def test_reload(self):
+    async def test_reload(self):
         key, init_value, new_value = "key", "value", "new_value"
 
         # put init value
         self.writer.put(key.encode(), f"s{init_value}".encode())
         self.finalize_cdb()
 
-        conf = Config.read(
-            filename=self.cdb_filename, reload_interval=1, loop=self.loop
-        )
+        conf = await Config.read(filename=self.cdb_filename, reload_interval=1)
 
         # put new value
         with open(self.cdb_filename, "wb") as f:
@@ -121,34 +117,32 @@ class ReadConfig(unittest.TestCase):
         self.assertEqual(conf.get(key), init_value)
 
         # wait until config file reloaded
-        self.loop.run_until_complete(asyncio.sleep(conf._reload_interval + 1))  # type: ignore
-        self.loop.run_until_complete(conf.shutdown())
+        await asyncio.sleep(conf._reload_interval + 1)
+        await conf.shutdown()
 
         self.assertEqual(conf.get(key), new_value)
 
-    def test_key_not_found(self):
+    async def test_key_not_found(self):
         self.finalize_cdb()
 
-        conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        conf = await Config.read(self.cdb_filename)
 
         with self.assertRaises(KeyError):
             conf.get("/missing_key")
 
 
-class ConvertYamlToCdb(unittest.TestCase):
+class ConvertYamlToCdb(unittest.IsolatedAsyncioTestCase):
     cdb_writer = cdblib.Writer
 
     def setUp(self):
         _, self.cdb_filename = tempfile.mkstemp()
         _, self.yaml_filename = tempfile.mkstemp()
-        self.loop = asyncio.new_event_loop()
 
     def tearDown(self):
         os.remove(self.cdb_filename)
         os.remove(self.yaml_filename)
-        self.loop.close()
 
-    def test_fill_cdb_with_yaml(self):
+    async def test_fill_cdb_with_yaml(self):
         _yaml = """
             service:
               db:
@@ -163,10 +157,10 @@ class ConvertYamlToCdb(unittest.TestCase):
         with open(self.yaml_filename, "w") as f:
             f.write(_yaml)
 
-        conf = Config(self.cdb_filename, loop=self.loop)
+        conf = Config(self.cdb_filename)
         conf.fill_from_yaml(self.yaml_filename)
 
-        cdb_conf = Config.read(self.cdb_filename, reload_interval=None, loop=self.loop)
+        cdb_conf = await Config.read(self.cdb_filename)
 
         self.assertEqual(cdb_conf.get("/service/db/connection/host"), "localhost")
         self.assertEqual(cdb_conf.get("/service/db/connection/port"), "5432")